Adds routes/campaigns.js
